[roll] Roll fuchsia [netstack3] Implicitly SNAT ports of non-NATed traffic
Now that we support masquerade NAT, it's possible for forwarded traffic
that is masqueraded to clash with locally-generated traffic. Because the
socket layer and conntrack keep track of available ports separately,
it's possible for conntrack to reserved some local port for a masquerade
flow that is forwarded by the stack, and then for a local socket to be
bound to the same port. When that socket sends locally-generated
traffic, it will be dropped at EGRESS when conntrack sees that inserting
it in the connection tracking table would result in a tuple clash with
the existing masqueraded flow.

To prevent this, the netstack implicitly performs source port remapping
if necessary to avoid a conflict. This should have no effect on most
traffic, but avoid a conflict like the above when it would otherwise
result in dropped packets.

The integration test added in this CL required some refactoring of the
SocketType trait that is used in filtering tests, in particular to split
its functionality into three phases: (1) bind, (2) connect, and (3)
send/receive. This allows tests to more easily reuse a socket after it
has been bound or connected.

Original-Fixed: 372543214
Test: netstack3-filter-test
      netstack-filter-fidl-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1133151
Original-Revision: fd2d3da7292c037249e4592753de80664e4cbab3
GitOrigin-RevId: 86c45d6253c8ed789f32bfd1ea042b5b4de0804b
Change-Id: Ie8bb98d31dad5c82261fac3fd29a9e3c0138ccca