[roll] Roll fuchsia [plat-config] add health check field to component configuration
The health check field represents a list of opted-in components
that can potentially block an update. It is currently part of
SWD codebase but is being redesigned to be in component framework.

The product sets an enum corresponding to a component that has
implemented the health check service and the enum is translated
to a component moniker to be used in the framework.

Original-Bug: 369128580

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1122712
Original-Revision: 4792f61c7e74db8c09d5e60adfcb18be73ca8595
GitOrigin-RevId: d317ea80850d46fde7eb48cb428baa1c574fd7ef
Change-Id: I0826656da7271ecfe343c3cdec32e6e3dc441553