[roll] Roll fuchsia [ffx] enable port:ip targets
Instead of requiring `ffx target add <target>`, just using a port:ip
should be enough for the daemon to try to reach the specified target.
(This functionality already existed, but only worked when discovery
was disabled. This change makes it so it is always used.)

Original-Bug: 354219263
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143503
Original-Revision: ac671802ffbd0882532675824a9dd9706d96fa15
GitOrigin-RevId: e94703045c80dc1b06182e71ccfdfb4aa2e1a47a
Change-Id: I0ae22375c837c7a4ae0f928187f34befe17e6dfe