[roll] Roll fuchsia [storage][conformance] Update Rights struct
The Rights struct in the storage conformance library should only be used
for Rights. Update it so that it is created with rights, and also add
functions so that it will return the Open3 right flags.

There were some conformance tests that did not run previously, probably
due to confusion of `valid_combos_with/without(flags)` which should not
be used for non-rights flags. Now that they are running, there were some
tests that fail and this CL also include fixes for them:

* stream reading past maximum buffer return OUT_OF_RANGE error
  (previously was returning INVALID_ARGS)
* the block device in the test harness for minfs does not have enough
  blocks and inodes
* truncate tests should only run if append is supported.

Original-Fixed: 293951181
Test: fx test storage_conformance_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143692
Original-Revision: 6ce53e406bde23a559bb60a9da0496ecfdc5a13e
GitOrigin-RevId: d1fa07c7e1a24a4e8eb4c59e8e9ea6e737e888c0
Change-Id: I9269464853bcff2877669482992c6bac902e1586