[roll] Roll fuchsia [gpio][core][test] Don't assume that pinimpl calls will be inlined
gpio-test currently uses RunUntilIdle() in a few cases that test that
the GPIO core driver calls out to the mock pinimpl driver. The mock
driver is on the environment dispatcher, so this is only guaranteed to
work if the calls are inlined. Switch to Run() instead of RunUntilIdle()
so that the tests will work regardless of how the FIDL calls are made.

Mutliply: fuchsia-pkg://fuchsia.com/gpio-test#meta/gpio-test.cm
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143348
Original-Revision: 1bf1d0b273876f5b00fb4d6b2065d93e0e8756ee
GitOrigin-RevId: 79c1c21f123146106894d4d6b1e925c5e5b9ef66
Change-Id: I6b8206fa72698e55eee01a4af902d6df0d2f4f5c