[roll] Roll fuchsia [fidl] Add type shape to external declarations
This unblocks Rust wire bindings, which need to know when external types
have out-of-line data. The type shape of a member would usually contain
enough information for this, but when external structs are optional the
field shape always contains out-of-line data (even when the struct
itself does not have out-of-line data).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1145253
Original-Revision: 9b1a0307d40ae905f8fabcfada7f608fe0a3ea87
GitOrigin-RevId: 93901872c801a481969100d350f115a5bbe8977e
Change-Id: I70c8f8926398210704af0324c41818d9fd8b0168