[roll] Roll fuchsia [fxtest] Fix dldist path handling
Finding the "dldist" matcher script did not respect the --dir passed to
fx, which caused several problems. First, if the script was not built in
the deal default directory, all matches would fail. Second, a failing
match showed as a failure to match, rather than a failure to run the
matching script itself.

This CL refactors a bit and uses a constructed path that respects the
directory argument. It also slightly simplifies the logic for selection
tests.

Original-Fixed: 375372016
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1146083
Original-Revision: 5c1d187042ea13157d2194d45153e2a203c02320
GitOrigin-RevId: 1246ad77e57316cdc4e001b80e8720ded2682d95
Change-Id: I5808702152349ef9091d6ce225f54cd7f3c0603e