[roll] Roll fuchsia [partition] Add Partition.GetMetadata
This returns all of the metadata for a partition that we'd expect to
find in a GPT.  For partition implementations that don't support some of
these fields, an absent value is returned.

This is redundant with the existing GetName/etc methods, but those have
been around forever and are too widely used to bother to port to this.

This will be used in the paver for preserving existing partitions across
a call to ResetPartitionTable.

Original-Bug: 371060853
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143900
Original-Revision: a73bf17d82543e2da9d8f2c76037a5ba9e5cdb17
GitOrigin-RevId: f673187f51ce49207d55a19d79579a2b518019d3
Change-Id: I7046740a308e31bc462288ab095614bba7a4e177