[roll] Roll fuchsia [tee] Defines TAs as `loadable_module()`s
This is rather than `shared_library()`'s, since TA's are meant to be
executables and not link inputs. This is of marginal build hygiene value
but also drops the perhaps-misleading prefix of "lib" from our compiled
TA filenames.

One consequence of doing s/shared_library/loadable_module is that we
lose the auto-generated .ifs link stubs created by our shared library
GN toolchain. This was recently realized to be of general diagnostic
value, as it gives insight into what subset of the TEE API is in used by
the TA. Accordingly, we begin to manually construct that in
fuchsia_ta_package() so as not to lose it. We also want this to
straightforwardly check at build-time that the official TA entrypoints
are indeed defined (to be done as a follow-on change).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143752
Original-Revision: c84b6ddeb03449cc73bd284c2b52c1d2965e322a
GitOrigin-RevId: 334dc45bdcbc8dde15a35e9e32b7a85a8cd45a59
Change-Id: I35c102ce594aea3f0e7b1ce8041e5ce185fe6a22