[roll] Roll fuchsia [debugger] Add JobOnly filter config for FIDL clients
Adds the ability for FIDL clients to specify JobOnly attaching, so that
they may be notified of exceptions in any children of a matching
component, for example, without taking the exception channel of all
processes under the job.

Since DebugAgentServer now needs to keep track of the filters it
installs, we can also share some of the code with the zxdb frontend to
turn FilterMatch structs into correctly configured AttachRequests with
duplicate matches overlaying settings as we expect.

Original-BUG: 362240229
Original-BUG: 362240382
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143812
Original-Revision: 9f67020a69a19af2895b4d67159f45105f3591cf
GitOrigin-RevId: 8470cceb859de55d27254f1895ebfbd5daaa1658
Change-Id: I40619a1ed41435b1bd5a0e8869a2c42a50f71205