[roll] Roll fuchsia [kernel][rppm] Clear pending power requests on domain changes.
Avoid sending a request intended for one power domain object to another
power domain when the request dispatch DPC races with changing the
power domain for the CPU.

The race condition is as follows:

1. A CPU is assigned domain D1.
2. A request R for a power level change becomes pending on the CPU,
   resulting in the power level DPC for the CPU being queued to dispatch
   the request.
3. Before the DPC acquires the scheduler queue lock for the CPU to latch
   R and D1, a different thread acquires the queue lock and replaces D1
   with D2 on the CPU.
4. The DPC acquires the queue lock and latches R and D2.

To prevent sending a request intended for the previous domain to the new
domain, we simply clear any pending request on the CPU when its domain
is changed. Since both the domain and request are protected by the queue
lock, the DPC cannot observe an inconsistent pairing.

Original-Bug: b/375221894
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1146697
Original-Revision: a7103f26efa67b6cfcb55a3374350178e79b9967
GitOrigin-RevId: 524e1557fcf3a34b76019ad84f091234c49cf920
Change-Id: I8bd65b18f371239528ba78d697b58a293aeb4ffd