[roll] Roll fuchsia [sdk][vfs] Disallow serving Service node types
The primary use case of the `Node::Serve` function is to serve *node*
protocols. Allowing Serve to be called directly on a Service node is a
bit of an anti-pattern for a few reasons:
  1. The resulting connection isn't managed by the VFS.
  2. Users creating a Service node already have access to the service.
  3. We cannot provide a type safe Serve since we cannot verify
     the actual protocol the service is serving.

In the event a node-only connection is required to the service, this can
be achieved by placing the Service in a pseudo-directory which owns it.
This is already required by every use case, since one cannot re-open a
service through a node reference connection.

Also change deprecated `Node::Serve` signature back to be a protected
method. It was accidentally made public in a previous change
(I48bfc161527f57ef53a336950fa939a3a2bf4c9f) allowing this functionality
to be accessed in nodes that don't support it.

Remove unnecessary deprecation notices.

Test: fx test vfs_cpp_tests
Original-Bug: 309685624
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1144776
Original-Revision: a4f80f0de2cca4edd5e8d186dfc47c682a76259a
GitOrigin-RevId: ddc4a7fb2abb8662c4bed9e5b67b08ba387c8e2e
Change-Id: I2edc629c528f88a036cb009d7206384b1fda59f7