[roll] Roll fuchsia [test] Prepare to support stable names for freeform metrics files
Return JSON | (str, JSON) from `process_freeform_metrics()` to allow callers in other repos to
tolerate both new and legacy behavior of `process_freeform_metrics()`

It turns out that the name of the file in which freeform metrics are
stored on test hosts is actually exposed in the database where the
metrics are stored. As such, dashboards might rely upon it. Using the
classname of the metrics processor, therefore, could cause outages in
the event of a refactor.

The most foolproof way to ensure that the author of a metrics processor
thinks about this issue and provides a stable identifier is to make the
filename a part of the return value from `process_freeform_metrics()`.

Original-Bug: b/376097015
Test: CQ
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1146197
Original-Revision: d6374681d9b4e4d3f4960e8652b9b67157f7ebae
GitOrigin-RevId: 90c7d0bca6b7eb4533ad1b8df407e3de221de8de
Change-Id: I61256e45c379d39bbae0892d32c261fff04da3f4