[roll] Roll fuchsia [netlink] Install routes in non-main tables
When a route needs to be installed in a non-main table, the netlink
worker now provisions a new fuchsia.net.routes.admin.RouteTable to hold
it.

Because rules aren't properly supported yet (see b/358649849), we still
need to write the routes to the main table as well.

This adds some complication. To cope with this, refactor things so that
the routes are held at rest as _FIDL_ route types rather than Netlink
ones. Because we're now installing routes in separate tables, the FIDL
Route structs that we observe through the watchers will now carry info
about which tables they're installed in, which aids us in making this
simplification.

This makes it easier to consume the FIDL routes watcher, as we can
always apply the watcher's events to our stored view of installed FIDL
routes, and then derive the netlink view of routes from that.

Original-Bug: 358649849
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1132574
Original-Revision: 184273343a7b495f3434b928c568090ef185cf62
GitOrigin-RevId: 037dd11c0602cc4d4ec98c8a41ed4ce0cef9837f
Change-Id: I8c1078e09de68f42f11f8051bc34945303d39d4d