[roll] Roll fuchsia [bazel_sdk] Override host tool paths when invoking ffx (again, again).
This is a reland of [1], which always includes the SDK manifest as
a runtime input, to avoid breaking OOT builds like the Android one.

Ensure that `ffx` and its plugins do not rely on the SDK manifest
to locate the host tool they may invoke at runtime, by passing their
location directly on the command line.

New functions are added to `ffx_tool.bzl` to retrieve the start
of the command-line invocation, that includes the necessary
`--config "sdk.override.<tool1>=<path1>,sdk.override.<tool2>=<path2>"`
argument.

This is required to allow moving the generation of the IDK to the
Bazel workspace. In this scenario, host tools will be built on-demand
as Bazel targets, and their location cannot be predicted before the
build, making it difficult to write them to the SDK manifest in advance.

(The SDK manifest will instead contain Bazel labels such as
`@fuchsia_idk//tools/x64:cmc` that `ffx` definitely cannot convert to
a proper host path on its own).

As two previous attempts at this CL had to be reverted, this has been
tested manually with an OOT Android Starnix-based build.

[1]: https://fuchsia-review.googlesource.com/c/fuchsia/+/1115539

Original-Bug: 367274537
Original-Bug: 287355615
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1123613
Original-Revision: 77a2250a9ddf54c0aad7ffb52af6af0c7f46392b
GitOrigin-RevId: d15e8a6b1c5c1f4a3642b8fa659ea27f02c0f7d7
Change-Id: Ib63842811d8d61a6c053ae5eeda1d1d271122093