[roll] Roll fuchsia [syslog] Remove syslog_internal sdk_component_manifests
This target, which was added in https://fxrev.dev/545048 and uses the
"internal" SDK category, is not in the SDK and all dependencies in that
CL have been removed. https://fxrev.dev/1121570 added a new target that
is in the "partner" IDK.

Remove the only two dependencies on this target. These dependencies are
on the directory, and this target happens to be the default target.

Original-Bug: 345827642, 372986936
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1144834
Original-Revision: 33881188f18b3ba7d99883674deb5a587007c221
GitOrigin-RevId: 8e690691b6a6acdbfe8b564ed3d9649d2e1a9454
Change-Id: I49bf7457e0f1bc53c2c43a8ce855638319a3459c