[roll] Roll fuchsia [assembly] Add support for bootfs pkg shell commands
Previously we only supported shell commands from base and full packages.
With this change, we also support shell commands that reside in bootfs
packages. Rather than place the trampolines for the commands in the
shell-commands base package, a new bootfs pacakge with the same name has
been added to house them.

Routing this directory to the serial console and ssh is still required
for those environments to be able to use commands in those environments.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1142872
Original-Revision: 3d3f365a96de7eeb81a57f29465de4741fc3eba6
GitOrigin-RevId: d6a1675bbb69264969c5adfb51b6a87d680ceb4f
Change-Id: Idd0b0429a7a796e6f736c20acbde60761de2404d