[roll] Roll fuchsia [shell] Route shell commands via shell-command packages
Rather than route /bin via the pkg-cache, it is now routed via the
pkg-shell-commands package's contents. Additionally, a new directory
boot-bin was added with the contents of boot-shell-commands. As
tools from /boot/bin are packaged, they will find themselves in that
directory instead.

This requires adding a the /boot-bin path to the global PATH so that
it is looked at for executables.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1144037
Original-Revision: ed67bdb52161fd6e5bd8bb1b6cc358bc6171e1bc
GitOrigin-RevId: 9ddc8f8fe4c88bce88d54e4ac20080a03fea64ce
Change-Id: I5fd4f196a034e5cf5b98650b0808ee30780247d4