[roll] Roll fuchsia [test] Support stable names for freeform metrics files
Return filename and metrics directly from `process_freeform_metrics()`

It turns out that the name of the file in which freeform metrics are
stored on test hosts is actually exposed in the database where the
metrics are stored. As such, dashboards might rely upon it. Using the
classname of the metrics processor, therefore, could cause outages in
the event of a refactor.

The most foolproof way to ensure that the author of a metrics processor
thinks about this issue and provides a stable identifier is to make the
filename a part of the return value from `process_freeform_metrics()`.

Original-Bug: b/376097015
Test: CQ
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1146979
Original-Revision: 3b0ecbc01b9734132bb2ec1fc7fe27695d1aded5
GitOrigin-RevId: ebaab3b79d6ab89ae5d728a2bd7d3264fc1deefd
Change-Id: I6bf9115ebc1227236cee722cabc53763ecdad5dc