[roll] Roll fuchsia [starnix][manager] Log which proxy wakes us
Log which of the proxied channels are waking us back up.
This makes it much easier to debug why the kernel is waking up.

Update the EventPairs to be in a HashSet so we can remove them
in O(1) without checking every Koid.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1146312
Original-Revision: eee287db504a45e271f24da429906933e6df67ab
GitOrigin-RevId: 434d5987001cf684bcd1c783617bcbb1c1e1a2bb
Change-Id: I3df9454597aa15c3228f82e83e0e7713a078f611