[roll] Roll fuchsia [netstack3] Skip implicit SNAT if conflict is identical
Currently if two identical packets end up creating exclusive connection
tracking entries and then race to finalize the entries in the conntrack
table, the second packet will see the first packet's connection in the
table and attempt to avoid a conflict by performing implicit SNAT. This
is not really what we want if the connection in the table and the
connection to be inserted are actually identical for all intents and
purposes; in this scenario, we want the second packet to simply adopt
the connection that's already inserted in the table.

Change the implicit SNAT logic to do this.

Original-Fixed: 375012282
Test: netstack3-filter-test
Multiply: netstack-socket-integration-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1145960
Original-Revision: e2ccd791bab0593a0c2e8e3baa0236dc1dbbb4c5
GitOrigin-RevId: 3030a2d67a8671ebd26aee75370d0e3faa6c5df9
Change-Id: Ie13f3cba52ae41613ba9a97a773d8c71814089e1