[roll] Roll fuchsia [wlan] Use Policy-provided BSS desc. during roam.
In a WLAN Policy-initiated roam attempt, the BSS info of the target BSS
is provided directly to Fullmac. This is similar to the way that the
regular Connect works. This CL changes the logic in brcmfmac to use the
Policy-provided target BSS info rather than asking the firmware for it.

Note that this CL removes a test case from the brcmfmac_sim_test
reassoc_test, because that failure case no longer applies once brcmfmac
is using the Policy-provided BSS info. This test also passes using
Antlion suite WlanPolicyInitiatedRoamTest.

Test: fx test brcmfmac_sim_test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1146574
Original-Revision: 336152d3bd10c31429c6f951c0d726de142b740f
GitOrigin-RevId: 63b39ef73e39614747ad5eee590305165058df7a
Change-Id: I164bd929b22442de19d2d3fa70d807f08112fd97