[roll] Roll fuchsia [ffx][injector] Make injector optional
This changes the type of the FhoEnvironment::injector to
be Option<Arc<dyn Injector>>. This is in preparation for having subtools
that do not interact with the daemon, or even do not interact with any
Fuchsia device.

The next step would be to make an execution path that creates a
FhoEnvironment with the injector == None.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1147592
Original-Revision: b531e5a73eebb6908b136f64c4e7caeb807f8c52
GitOrigin-RevId: 72967f0fa3affc7782d6531635825d2501cca316
Change-Id: Ibdff5ceaead6f87924f679f450aec3daac390369