[roll] Roll fuchsia [logs] Route LogSink to gtest tests
Route LogSink to gtest tests.
This isn't normal routing, and just forwards
the entire /svc directory, which is a temporary
workaround for logs not working in this context.
The workaround has the additional consequence of
logs being misattributed to the runner instead
of the component that is running.

Original-Bug: b/376735013
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1148816
Original-Revision: e4a08fd08ed2c88bbb9d9a0c123a5af914e5cdc8
GitOrigin-RevId: 7f70a69fcb8e54cde8c07c468efb6882a5e8e2eb
Change-Id: If843c28c3d3cb69db4d8c85450645f84bfd84498