[roll] Roll fuchsia [packet-formats] Simplify IPv4 options
We only need an enum for IPv4 options, and we need not encode length +
data on unrecognized options, since those are redundant.

Introduce a `copied` fn instead of the `copied` value since the `copied`
property is mandatory for router alert and we can see the kind bind as a
single unit instead of parsing separately into a boolean.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1149362
Original-Revision: 86d6bb6218a48484b053a8e0e21df205794920ed
GitOrigin-RevId: 81e3e9b93b825949081fbeeace87aed78774ee78
Change-Id: I14b3b0b7bf0463ba4ce05b15732807f340121713