[roll] Roll fuchsia [sestarnix] Introduce AccessDecision to query APIs
AccessDecision encapsulates the allowed, audit-allow and audit-deny
bits for each permission, and a flags field which is currently used
to indicate whether or not the source domain is listed in the
policy's permissive domains map.

SecurityServer permission-check tests are updated to validate that
the expected audit decisions are made, and test cases are added
for the handling of enforcing, domain-permissive and global-
permissive access checks against kernel classes/permissions that
are not defined in the policy.

Original-Bug: 362706116,
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1147712
Original-Revision: 165c61b6b94f02ab68f4b5abb583300f1d608a85
GitOrigin-RevId: 4a326b7360e2fa74389ad306f01532c1a1a42eaa
Change-Id: Ic28a27968775276171dd6a8613b2142bdee7a066