[roll] Roll fuchsia [fuchsia-async] Reimplement TaskGroup in terms of Scope
As a next step we can remove TaskGroup entirely, but I wanted to swap
out the backend as a canary for implementation issues and so we can
announce zxdb support for scopes.

This required adding ScopeRef::spawn_local and while we're at it,
ScopeRef::compute_local APIs.

This removes the TaskGroup::add method since it wouldn't be very
efficient to implement. We might want a reparent method eventually but
shouldn't encourage it as a way to spawn tasks.

Original-Bug: 339724492
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1148332
Original-Revision: 4da2f8dbad54ce04142b98293e94fa1392320946
GitOrigin-RevId: a543a3fed14523e95c84120d2bae71efafe30193
Change-Id: I4afb2933a206f02dde582d902aff75097f54cb3a