[roll] Roll fuchsia [sdk] Revert zircon-internal to a standard source_set()
https://fxrev.dev/979816 converted "zircon-internal" to an
`sdk_source_set()` because it was a `public_deps` of "fake-object", but
it is not actually used by "fake-object".

This reverts changes to zircon-internal in commit f6a44a00adb4da2219f225456d91adbb8381c9f6.

Remove that dependency and correct dependencies in "devicetree" so that
there is not an (SDK) public dependency from that SDK static library.

Original-Bug: 372986936
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1149581
Original-Revision: bdacd65195dfe310cee029f97b15098302c3f31d
GitOrigin-RevId: 08495e11942fba4a8d3d5a61be809e90744e3e06
Change-Id: I07824dab5e374f34a487ee288a76fbae114d2e48