[roll] Roll fuchsia [sestarnix] Set xattr on FsNode, if supported, on creation
When a new node (file, directory, symlink, etc.) is created in a
filesystem, an in-memory label is determined for it based on the
labeling scheme configured by the loaded policy.  If the filesystem
uses "fs_use_xattr" labeling then the label is also serialized and
returned to the Starnix VFS, to write to the new file node.

While in principle no xattr should be returned unless the
filesystem is configured with "fs_use_xattr" labeling, use of the
"fscreate" task attribute when creating files in non-xattr-capable
filesystems requires that ENOTSUP is silently ignored, for now.

Original-Bug: 355180447, 376429620
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1150336
Original-Revision: 24a04c4484ad6c7f5ca57e0c26ecd7d2da49e828
GitOrigin-RevId: 88b0d48f56f4286032e522598a7ebc772600d195
Change-Id: I04f83781ce95de160ef41830d7b618c3aeebfd82