[roll] Roll fuchsia [symbolizer] Fix Negative Loads with Multiple Modules
Zircon on x64 has negative loads, which is accommodated by treating its
base address as 0. This caused a conflict when searching through the
modules looking for an address by upper/lower bound because it would
find a different module and find that the zircon address is not
contained in the module.

This works for symbolizing kernel backtraces right now because in every
instance, the zircon module info is the only info that's provided, so
the upper/lower bound just happens to find it.

This patch adds a fallback to check for a module with 0 base address if
the wrong module is found.

Original-Bug: 377355351

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1150712
Original-Revision: 8f57a06ddc606d741851bee8d4e9cd197c7ed4e8
GitOrigin-RevId: 2c762b48008d20afd8373a34c0aa0f9c8c6b4a98
Change-Id: Id20de074e9a27a6d9b9975856f242854962c48b9