[roll] Roll fuchsia [build] Shift large bootfs rust binaries to larger RBE workers
This shifts a number of larger, slow-to-compile, rust bootfs binaries
from the 2-core to the 8-core RBE workers.  This reduces the
compilation times (using RBE) for the following (vim3-debug, which uses
1 codegen unit with thinlto):

- component_manager: 8min30sec to 4m30
- fxfs: 4m20 to 1m40
- pkg-cache: 3m45 to 2m10
- shared-multi-call-tool-binary: 3m20 to 1m50
- driver_index: 3m00 to 1m50
- fshost: 2m20 to 1m55 (note 1)

[1] - This particular target didn't benefit much with this CL, but the
next CL changes vim3 to use the same 16 codegen units that the other
arm64 targets use, and it was 3m5s with that change, so I added to this
set to keep them all together.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1151112
Original-Revision: 9a5e4e8f644ec21fd87612df2540b18943ef5281
GitOrigin-RevId: df9a5b1501fd297cd2a020d70fb5b586f9018a14
Change-Id: I8d7f8549d73b5dce37468d7671d35d1043cb92dc