[roll] Roll fuchsia [selectors] Fix validation of string patterns
We were requiring some chars to be escaped when it was unnecessary for
htem to be escaped. The only thing that should require escaping in a
string pattern is the `*` when used as `**`.

Escaping of `:` and `\` is necessary at the selector language level, but
it's not necessary when we are just dealing with string patterns. We can
still escape `*` in string patterns, as that has a different meaning
than a raw `*`.

This is similar to how we unescape exact matches. We should have been
unescaping string patterns as well, taking into account the edge case of
`**`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1151239
Original-Revision: 7b0257db3770f0ec72a4d89ffd573a9ccc8a2f28
GitOrigin-RevId: bc181aeb3316f5ee02ab40bf529eb397b9e386c5
Change-Id: I8803bc276f0f7f92153b21716215ffaf33e82b0b