[roll] Roll fuchsia [inspect] track "at capacity" duration of history
This patch adds a specific "duration of history"
stat for when the buffer is running at capacity.

The earlier "history_duration_ns" is useful for
humans but because it also measures duration in
the early phase of the buffer, it interferes with
Cobalt MIN_PROCEDURE sampling. This variant is
vended only when the buffer is at capacity.

Also, adjust inspect type to "int" to better match
the zx duration type.

Original-Bug: 365600132
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1151036
Original-Revision: a45945d7838ad02b0bbe1a68c37d9f644a892864
GitOrigin-RevId: 0bffda79ccea8386bfaf2cc9bc9954349e8e5403
Change-Id: Iaa27b008ea2e6bd6ebefeebc232cf8d069ecd78a