[roll] Roll fuchsia [paver] properly report last-boot-attempt
The firmware sets a slot to "unbootable" just before the final boot
attempt, but this caused some unexpected behaviors in the OS when the
current boot slot is reporting itself to be unbootable.

This CL adds detection for this case, and adjusts the paver FIDL data to
properly report the slot as pending but with zero attempts left.

Original-Bug: b/368597963
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1148338
Original-Revision: a7e5c7b8d30f25f9ce316b657b20b93876ecbf30
GitOrigin-RevId: cb55c1e15d2309594624f971d3e17690af3bf0e2
Change-Id: Ic860a8e3b340742fd481d784526574863a0b2eb9