[roll] Roll fuchsia [display] Move utils to standalone library.
This CL does not introduce any new code. All utilities are moved from
util/ directories in existing drivers. The drivers themselves are
modified to use the headers and classes/functions in the library.

Updating driver usage includes adding the display:: namespace qualifier
to the classes and functions that are now in the library. This accounts
for the vast majority of the driver changes.

Original-Bug: 359960626
Test: fx test //src/graphics/display/lib/driver-utils
Test: fx test //src/graphics/display/drivers
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1151373
Original-Revision: 565e443d67a8b7b61b1ee7f8e752be329a358d9e
GitOrigin-RevId: ae82041f547673fd2413158711fad18353be6a2f
Change-Id: Iaffab6a03946207c18cf8f0ae24ba4fb86a11d98