[roll] Roll fuchsia Reland "[docsgen] Fixing clidoc dependencies"
This is a reland of commit 87ebfa6dcee9f19e8a32e6d0ace1a4c40cd4f128
The revert was due to a doc change that was a problem which slipped in,
so unrelated to this CL.

This uses the sdk_molecule manifest as input to determine
the list of tools to run to generate the reference docs.

This is done to make sure all the tools are dependencies of
the invoke_clidoc target to make sure the build is consistent.

Original-Bug: 376723774
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1152353
Original-Revision: e185650e57ac569c159998f7790f67ff327e3aab
GitOrigin-RevId: 885e7b1ad690eb7186b18576b8ff1baada0cca5f
Change-Id: Iab0cceb6ec3eb804bdf7b87376f691677c5a60cd