[roll] Roll fuchsia [sestarnix] Add permissions checks to fs_node_setsecurity()
When a file node's "security.selinux" xattr is modified, the system
validates that the task has "relabelfrom" and "relabelto" rights to
the new and old Security Contexts, respectively.

The new Security Context should also be checked for the "associate"
permission to the file-system; this check is made, but the result
ignored, until correct file-system labels are implemented.

Original-Bug: 351195217, 355180447
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1112132
Original-Revision: ef955fef3cb849bd17334b626d72b801dcb3e13b
GitOrigin-RevId: f5267470dabeccd13557f114247e3e4959d7a323
Change-Id: Ia08a82cf39ad65f34755c6ec0080eefb308b7af4