[roll] Roll fuchsia [session] Remove builtin runner + boot resolver session-env
We route some capabilities to the session that it does not use, and are
only used to launch platform components. To protect against an unwanted
dependency in the future, remove them from the session's environment.

These capabilities are:

- boot-resolver: For launching fuchsia-boot components
- builtin_runner: For launching builtin components (components in
  component_manager). An example is the elf runner component.

In the case of #2, we are adding more builtin components, which were
designed only for platform use.

Original-Fixed: 375000535
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1146981
Original-Revision: 9f24dd261eb63855986e877acf7bfd83bd1c1cc7
GitOrigin-RevId: 24078a28b798d8836085b3ea247689e78b97d830
Change-Id: I1746785619982cbebd280dc237b80d0661b162c8