[roll] Roll fuchsia [build] Enable frame pointers by default in release.
The absence of frame pointers in release builds is the only
obstacle to being able to run performance profiling tools like
the CPU profiler or heapdump on "regular" release builds. It
would streamline workflows and reduce the number of bot
configurations we need to be able to use the default release
artifacts for profiling.

This does increase the size of the system image, but it is a
dependency of the project(s) to remove unwind tables which should
eventually net us a very significant binary size reduction.

Previous toolchain investigations have shown that the performance
overhead of frame pointers is negligible.

Bumps several core size limits to ensure that there is at least
as much headroom as a single instance of each bucket's creep
budget.

Original-Bug: 42164648
Original-Bug: 42053657
Original-Bug: 369225659
Original-Bug: 359658592

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1147933
Size-Review: Aaron Wood <aaronwood@google.com>
Original-Revision: 8b5f42ec2aecbbd494ab6211f25fc2139491bc50
GitOrigin-RevId: eed359f5827e5f97d3eff862ad8eaa830567232a
Change-Id: I7b6e7e142ef9adb8249c570e92cc1ed9890b6659