[roll] Roll fuchsia [sdio] Remove special case for Broadcom SDIO chips
Remove a check for Broadcom chips in the SDIO core driver that assumed
that function 0 was included in OCR. It turns out that this may not be
the case, as the third I/O function on Nelson appears to be valid.
Having an extra function node should not be a problem anyway because
brcmfmac is already configured to bind only against the first two.

Original-Bug: 376074123
Test: Booted Nelson, used sdio utility to dump info for all three SDIO
      functions
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1153313
Original-Revision: 463187b8e155682d3208cfa30459d10a3cc2b8d2
GitOrigin-RevId: 0a2695d5e6506c863d35424219f9417f1012da59
Change-Id: Iac9afac151dfa83e260e143da971f5f0e125ab9e