[roll] Roll fuchsia [fuchsia-async] Rename ScopeRef to ScopeHandle
And document the scope lifecycle better.

I haven't liked the "Ref" naming that much, since it evokes the idea of
&T, so I prefer "Handle".

JoinHandle is kept the same because of its similarity with the ones in
tokio and in std::thread. It is much more limited in its capabilities
compared to a ScopeHandle, so renaming it to TaskHandle by analogy to
ScopeHandle (as this CL originally did) suggests more similarity than
there actually is.

NO_IFTTT=ScopeRef does not appear by name in zxdb source

Original-Bug: 339724492
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1152852
Original-Revision: cd97d8285eff82c4ded148a7cbe80d696391d0b2
GitOrigin-RevId: afff01c29c9b32c0aa9889a0b95278b2ab461b36
Change-Id: I629dd2b775d5b962cd3b21606f082bce4adda6aa