[roll] Roll fuchsia [logs] Fix error with interest setting when reconnecting to a device
Previously, when a device rebooted, if an interest level
was specified, the reconnect would fail due to the proxy being
closed, because the proxy was passed outside the general connection
object. Now, the proxy is part of the connection object so
this should no longer happen.

Original-Fixed: 368353859
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1153818
Original-Revision: cde44e9bd58450870269a5cabea8632299a3cf13
GitOrigin-RevId: ea9c1f6c9a2eff1efc81ec88e77990d9d53b3d54
Change-Id: I31ceed671eebf82e044f7afe813fe944bc7f6075