[roll] Roll fuchsia [netstack] Make address properties type safe
Replace custom validation with defined types for a positive monotonic
instant. This leaves us in a better place to eventually migrate these
fields to the boot timeline, now that the types are explicit about it.

For now, conveninence translations from the monotonic timeline (when
building on Fuchsia) are available.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1153812
Original-Revision: 8fc08486d7fc0c37d8cdf309fbdeb68b82beab01
GitOrigin-RevId: 4ec0bf45a02e77b42e7033ea77d518052ece33e4
Change-Id: I11f346046869dfb122374e88c5d5af18c39df65d