[roll] Roll fuchsia [network/bench] Read iPerf results from file instead of logs
Have iPerf write its results to a file in /custom_artifacts rather than
attempting to retrieve them from syslogs, as it's possible for Archivist
to either process the log lines too late for the benchmark to observe
them or to drop them altogether.

This requires adding a custom proxy for custom_artifacts to the netemul
sandbox, that proxies the custom_artifacts storage capability as a
directory to components in managed realms. This is necessary because
storage is deleted when a component is destroyed, which means any custom
artifacts written by a test component are deleted when the test realm is
torn down. The netemul sandbox creates a new subdirectory for every
component to which it routes custom_artifacts to provide the
per-component isolation that it would otherwise get out of the box from
using a storage capability.

Original-Bug: 378163044
Original-Fixed: 377159657
Test: iperf-test
      netstack_iperf_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1152492
Original-Revision: 0b390e4734c0ae942923fac14cecbbaeb5488b13
GitOrigin-RevId: 578228106a329eaec104ad6b9a32c2bf8ba02bbd
Change-Id: I41ec3633454df598c3addb861391c9100d678a55