[roll] Roll fuchsia [flatland] Temporarily disable direct-to-display.
fxb/370839049 aims to relax constraints on image usage by the display
coordinator.  As an intermediate step (fxb/370994314) we want to remove
the abililty of the coordinator to signal release fences directly.

Typically this involves migrating from observing event signals to vsync
stamps, as was done with Carnelian and the Vulkan swapchain.  The same
could have been done here, except that unlike the other cases it would
be throwaway code (since, here, events are used to enforce the
constraints that will soon be relaxed when fxb/370839049 is fixed).

So instead, temporarily disable direct-to-display, and rip out the code
that relies on events.

Original-Bug: 377979329
Original-Fixed: 377967775
Original-Fixed: 42173333
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1153340
Original-Revision: ec5229c6ffe4fa9e53ef335648f222fedecfbaad
GitOrigin-RevId: a99d08c2526c1ef28ec64ba8834ff2d9ae4eb7ab
Change-Id: I63eb3193f35190febcb25ed76bf064e85657c767