[roll] Roll fuchsia [nmfs] Add metric for no events to replay
When the filesystem attempts to reconnect to the socketproxy and there
are no events to replay, such as when the last present network has been
removed, these cases were being incorrectly logged as
"successful_reconnects". We cannot be confident that the reconnect
was successful unless we make a call against the proxy such as
Add/Update/Remove network or SetDefault.

Extract this case into a separate inspect node "no_events_to_replay" to
separate this from being classified as a successful or unsuccessful
reconnect.

Added test to confirm this behavior. This test is separate from the
existing test that validates multi-threaded functionality to assist with
debugging in the case of a while loop hanging.

Original-Fixed: 375233401
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1144798
Original-Revision: 3086cc9b1de729a186450445fe77eba4f28c8c97
GitOrigin-RevId: abae2a3cc08f135ca0cd45e4ab3d8e76b8f65c6d
Change-Id: Ie98d4e53f336b4a35e6b8982e2c9bfd663b3cee4