[roll] Roll fuchsia [starnix] Carry kmsg log levels to fuchsia syslog severity
Instead of just dumping the message as received, we can carry over the
severity and map to a fuchsia syslog severity. For messages without a
direct mapping we can continue printing the exact level that we received
but map it to a reasonable severity. For example, NOTICE can be an INFO
but we still print <5>.

Original-Bug: 378507649
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1153772
Original-Revision: 8ccee5298089c0d30841efa7af51bacad0cebf50
GitOrigin-RevId: b9529cebe84497303e4b9f1be5cb263037b4e306
Change-Id: I8bc54bc2509e8d424a5fede516308a5ab13b163c