[roll] Roll fuchsia [fuchsia-async] Add close operation to Scope
This is needed in archivist and inspect and probably other places for
handling system shutdown and test scenarios. The basic use case is if
you need to begin graceful shutdown and expect the tasks in your scope
to follow suit because of some coordination happening externally.

Original-Bug: 339724492
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1154862
Original-Revision: 0215619f537c67de115717443a28736f4cdfb647
GitOrigin-RevId: 0295e7cbac0fa616d3c98b138fe86ad1557c6888
Change-Id: I23b153a2c6ab87fb6b84ec4c3afc893adafc2394