[roll] Roll fuchsia [kernel][ktl] Update __config_site for upstream protocol changes
The libc++ __config_site file is now expected to define various
macros to either 0 or 1, where previously it was only expected to
define them (e.g. to empty) or not at all.  The standard behavior
of the preprocessor (without -Wundef) means that not defining
something at all is treated as defining it to 0.  So this omits
any explicit 0 definitions, which older headers using #ifdef
rather than #if would treat differently; but it updates each
thing actually defined to be defined to 1 instead of empty.

The libc++ headers have also changed to require a few more things
to be explicitly defined that were previously implicitly derived
from other settings.

Original-Fixed: 378145866
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1155175
Original-Revision: c28845b0f034dd613992fd895cb121da8ce3e52a
GitOrigin-RevId: 67bb8b3f01cecf67edc868d843ea316063a4c05c
Change-Id: I2c08b49a759b1b526ffdc9732ab76db5e6adf623