[roll] Roll fuchsia [fdf] Add support for stable bus topology info
This change allows bus drivers to supply bus information when they
create nodes, as well as provide information about the stability of the
bus nodes it provides. Lastly it provides a way for drivers to expose a
token which can be exchanged but the driver's clients to determine what
the bus topology for these drivers should be.

The bus topology can be reasoned about to such that clients to
reasonably identify the devices they are talking to and ensure it won't
change. Because it is structured, clients also will not need to parse
the bus topology with things like a regex to understand the
underlying bus topology for the device.

Lastly, tokens are used to prevent drivers from both learning about the
the bus topology as well as ensure the client can trust the bus
topology.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1148812
Original-Revision: 9b41af0e3fc06eeb4cb8016e4a8bef6ab04e2b15
GitOrigin-RevId: ffd9555950f22c378b2d879993d27b561c8cd156
Change-Id: Iccf5456235eb6349542bbf39cd5dd23b6ee13736