[roll] Roll fuchsia [archivist] Simplify shutdown logic
I suspect a bug in this logic was the root cause of an old rare flake
in Archivist's Stop integration tests. It could have been the case that
the `logs_repo.wait_for_termination` future was resolving *before* we
got the stop signal which would lead to not ingesting logs/inspect and
that test flaking. I realized this in the follow-up when I was
observing tests consistently failing due to this.

Multiply: archivist_integration_test
Multiply: archivist-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1155572
Original-Revision: d1ba5741498863824068d3b76e0bdb089b1ee140
GitOrigin-RevId: 38511893ff3fa2d525f15cac862b9d8d39b9cb8a
Change-Id: I59262c27614d865015e4137eaf2ee06a0d28a38d