[roll] Roll fuchsia [sandbox][components] support update notifiers being registered on dicts
Adds support for update notifiers to be registered on sandbox
dictionaries. These are functions which will be invoked whenever an
entry is added to or removed from the dictionary on which it's
registered.

Also uses this change to keep VFS directories created from a dictionary
up-to-date when the contents of the dictionary changes.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136134
Original-Revision: 6e39679c41c6dca65a537a8bad15b54d1646f362
GitOrigin-RevId: 29d4c385eee2295de585ceb0c2cf0bf25db086fa
Change-Id: I1fb97ba4a11d9214bb8f65c7ef0e0aeeeaab3a94