[roll] Roll fuchsia [rbe][cxx] Set remote compile timeout 10m
The default timeout is 1 hour, which is way too long to
wait for a timeout for compiling.
Reduce to 10 minutes for compiling, which is already
more generous than it should be.
(Reduce further later, if deemed appropriate.)

Linking is not affected in this change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1156573
Original-Revision: b7213d47d81c82042a65247899992955965741e3
GitOrigin-RevId: c4f259b29533f7c7f12c9b89506f39989febf623
Change-Id: I7829cc9cd0084e78fec138ec164dc4b456468b57