[roll] Roll fuchsia [rust][fidl] Bits support for new Rust bindings
This change also refactors the way constant values are emitted to
support bits more comprehensively. One of the compound identifiers
included in the IR turned out to be a more specialized subset of
compound identifiers, and may include a member path as well. This wasn't
as much of a problem prior to bits support.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1144777
Original-Revision: 93ad9bf4320e3189d6f95c338e73084bd2b4c219
GitOrigin-RevId: 6631f95d113558578490714f84b25a629e55ddd7
Change-Id: I849aeff2cfe81cf50e214a109ed48243f38281f8