[roll] Roll third_party/pigweed pw_bluetooth: Fix RFCOMM information length off by one
A misreading of the spec led to confusion on whether credits was counted
as part of information_length. It is _not_, but it is included in the
calculation for max allowed payload.

Fix this and standardize on `information` as the name used in the spec
vs payload.

Also add helper flag for checking whether a frame is UIH or not

Add standalone unit tests for RFCOMM.

Original-Bug: 378692444
Original-Bug: 378691959
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/249092
Original-Revision: 40d4c1261275065d30e77ea68fbea69061b70e78
GitOrigin-RevId: 13b81cd5ce5bf308b04663d15970aee06e6dc367
Change-Id: Ic793b6a7c697f83cc9204dca2c18f687be6bf3e1