[roll] Roll llvm-project/libc [libc] Remove more libc dependencies from the RPC header (#116437)
Summary:
The end goal is to make `rpc.h` a standalone header so that other
projects can include it without leaking `libc` internals. I'm trying to
replace stuff slowly before pulling it out all at once to reduce the
size of the changes.

This patch removes the atomic and a few sparse dependencies. Now we
mostly rely on the GPU utils, the sleep function, optional, and the
type traits. I'll clean these up in future patches. This removed the old
stuff I had around the memcpy, but I think that it's not quite as bad as
it once was, as it removes a branch and only uses a few extra VGPRs
since I believe the builtin memcpy was improved for AMD.

GitOrigin-RevId: 2c4b32342c550396e95f8a5933b50fa11d782afa
Original-Revision: a18016fec582d0e428a091f1401bc02de89246a9
Change-Id: Ib0f1abdf5367f50596101fda0be21371b0cfbbc1