[roll] Roll fuchsia [superproject] Roll llvm-project/libc [libc] Remove more libc dependencies from the RPC header (#116437)
Summary:
The end goal is to make `rpc.h` a standalone header so that other
projects can include it without leaking `libc` internals. I'm trying to
replace stuff slowly before pulling it out all at once to reduce the
size of the changes.

This patch removes the atomic and a few sparse dependencies. Now we
mostly rely on the GPU utils, the sleep function, optional, and the
type traits. I'll clean these up in future patches. This removed the old
stuff I had around the memcpy, but I think that it's not quite as bad as
it once was, as it removes a branch and only uses a few extra VGPRs
since I believe the builtin memcpy was improved for AMD.

GitOrigin-RevId: 69c0118a72768c1d1ef2395668df96d20cf93841
Original-Revision: a18016fec582d0e428a091f1401bc02de89246a9
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1159132
Original-Revision: 7669040cee52a7d071e28f7a9b176c17a930e5a4
Change-Id: Ibfc8c3f210b3a458071f8137769fa94e5692093c