[roll] Roll fuchsia [factory] Migrate Clone -> Clone2
Node.Clone is being replaced by Node.Clone2 as part of the ongoing io2
migration. The latter method is already supported by all in and out of
tree servers at all API levels.

All usages in src/factory are in tests, so we clone all connections with
the original rights instead of ensuring they are readable. This can be
resolved in the future by re-opening the directory with the dot-path if
required.

Original-Bug: 378924331
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1155955
Original-Revision: 48b5f223664bd8bdbd752a86ee1d9451a1903eed
GitOrigin-RevId: fdf44ef5a273528d5a02c51c3dbe8978a327c958
Change-Id: Iaf4c642b88f259bca66e31b05f4d2f3c367f399b