[roll] Roll fuchsia [kernel] s/__code_start/__executable_start where appropriate
This change goes throughout kernel source and replaces the usages of
__code_start that represent "start of the executable" to
__executable_start: these symbols coincide when the kernel is linked
with kernel.ld and loaded as a ZBI, but differ in the elf-kernel.ld +
ELF loading context, where __code_start actually refers to the beginning
of the .text section. This change should be a no-op, and ultimately fix
the main bug preventing loading the arm kernel as an ELF (this
discrepancy causing pathologies at runtime).

Original-Bug: 42164859
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1154774
Original-Revision: 111586c44449a10582cbbba2ba18e8e7fd63e8b3
GitOrigin-RevId: 646086feda74a3c61a833417410a1dce13b81a7e
Change-Id: Ibdfb819d2cbd83510d8bceb34a0fcdb5ff11aa8d