[roll] Roll fuchsia [botanist][build] Add virtual_device_spec to environment.
In order for botanist to use ffx emu with the product bundle, it needs
to know which virtual_device spec to run with that corresponds to the
desired emu_dimensions. This change also modifies the existing device
specs to match the 1cpu environment and infra default values. Botanist
will only switch to using the product bundle if the recipes pass the
virtual device spec through to botanist.

Original-Bug: 377332712
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1153900
Original-Revision: a94689b19abda3894d7ed3a9a9e74677618cc38a
GitOrigin-RevId: 4a342d6a29b3ec42d04ac69ae619a8bc6e7be980
Change-Id: I17f6067c6b02c071b7b6ccfc93b1b32ae8ac6fb2