[roll] Roll fuchsia [owners] Remove jyotiraju@google.com from sdk/fidl/fuchsia.telephony.ril/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/jyotiraju@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1160123
Original-Revision: 2d3649b56f495ce7b83fc8ef357815f523ebb0b8
GitOrigin-RevId: 77cfc4bbced8d523eae29c102afdfb5f24a70b8f
Change-Id: I19d2c2bf3f0e5d726956f9d1dd721de4ad47ee4f