[roll] Roll fuchsia [owners] Remove jyotiraju@google.com from src/connectivity/network/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/jyotiraju@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1160124
Original-Revision: ef2836074a69d4eea83071d989227bfa5d758d35
GitOrigin-RevId: b632ec67994e4a981bbfdf514e8ddfe057db9775
Change-Id: I1391ed1607dd9a8973a0a5985a55d96b78ca7e3a