[roll] Roll fuchsia [tee] Implement binder calls to connect to TA and invoke a command
This adds binder protocols for connecting to a TA and setting up an RPC
session bound to a session with that TA. virtio-socket ports are
allocated sequentially based on the order they appear in the
configuration exposed by the TA manager. In this change communication
with the TAs themselves is stubbed out.

Original-Bug: 369435662
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1129756
Original-Revision: 7d9bc7dc3d80dfdbb391fcaf4d176479c4fd42bf
GitOrigin-RevId: 19da17589f5dc7ad913b6859f314e8d88e922c09
Change-Id: I59a185891f3fce593b175823b193cb9d6750ed22