[roll] Roll fuchsia [rust] Migrate off of async-io
This migrates our async framework off of async-io
and onto Tokio. Unfortunately, Tokio timers don't work
the same way as Fuchsia timers in the sense that an
executor is required to create them on host,
but on target you don't need an executor to create a timer.
To workaround this limitation, the timer is lazily initialized.

Original-Bug: 378000076
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1159422
Original-Revision: 767d21e50df909d94d1715e56eac1ee7e80de4cb
GitOrigin-RevId: be7fb8241c15d4115238989d57996c72a61e1875
Change-Id: Ice3baa2aae35e211f02cf3f28a16002dcfc779b7