[roll] Roll fuchsia [component_manager] Support tracing when nested
At startup, nested component_manager will connect to a trace_provider in
its namespace, if it has one. As a proof of concept, the
fuchsia.component/Controller.Start method is now traced. You should be
able to get this trace if you run `component_lifecycle_benchmark-test`
with `ffx trace`, for example.

To support this, component_manager has been branched into two variants,
`//src/sys/component_manager:bin` and `:bin_eng`. The only difference is
that `bin_eng` links in libtrace_provider and enables tracing. Splitting
the targets in this way was necessary because linking libtrace_provider
adds ~1.5M to the binary (due to linking shared libs, it seems like),
which is not a tolerable cost in production. For now, `bin_eng` is only
used in nested component managers.

In the future, we may use `:bin_eng` in eng builds, and support using a
trace provider from a child (it won't be available in the root component
manager's namespace). This will require some assembly plumbing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1147614
Original-Revision: 57f731d32449c41d1b60a939d4a9fd589d123990
GitOrigin-RevId: 17f374fedb7538a1f3c3e05c6e8d623f57e010b5
Change-Id: I02173bc8c742bc274ab81a69486af2ef9e00f701