[roll] Roll fuchsia [kernel][phys] Disentangle use_elf_kernel from kernel.phys.next
This change removes the connection between the value of use_elf_kernel
and what value of kernel.phys.next is set.

The way the build integration is currently organized, setting
use_elf_kernel = true results in setting an additional
kernel.phys.next option. We remove this connection and instead have each
of the ZBI and ELF flavors of kernel_elf_binary() override the
distribution_entries to ensure they get packaged as "physboot", which is
the default expected value of kernel.phys.next. This sets the stage for
when we'll eventually delete the ZBI one and rename the ELF one as just
"physboot". Moreover, this conveniently sidesteps an internal hoop that
we'd have to jump through when throwing the use_elf_kernel = true switch
for arm relating to internal CMDLINE-diffing logic around arm products.

Original-Bug: 42164859
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1157393
Original-Revision: df7fcc31a4aeed95396f4f70aac1ff297149a21b
GitOrigin-RevId: 2c4f76cc08f013fa16f74554412bdb741608b916
Change-Id: I801ee216ce4e2e437f5aec505c6520aad70731ff