[roll] Roll fuchsia [network/tests] Deflake neigh_clear_entries
The test is flaky because during the first round of UDP exchange,
Alice may send multiple NDP solicitations, and Bob will respond each
of them. The race happens between the arrival of Bob's advertisements
and the removal of Alice's NDP entry. If one of the responses arrives
later than the removal, then it will install a new entry, causing the
test to fail.

This change relaxes the test so that when it detects a race, it will
retry up to 5 times.

Original-Bug: 379379169
Multiply: netstack-neighbor-fidl-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1159912
Original-Revision: 1f6799969fb600f0ce8846563d0970b722be2329
GitOrigin-RevId: efb6c57b3597ed14293756de0ba28f2a6a81adae
Change-Id: I23c7e134f1451325ae4a7d122981af26f464b3ff