[roll] Roll fuchsia [display][api-types-cpp] Rework AlphaMode.
This CL updates AlphaMode to be its own type (as opposed to an alias for
a FIDL type), so driver code is guaranteed to use IsValid() and ToFidl()
/ ToBanjo(). The update gives AlphaMode constexpr conversion methods, so
it can be used in higher-level types that offer constexpr conversion
to/from FIDL/Banjo.

This CL also calls out that we haven't updated the AlphaMode interface
for API stabilization, so it's not suitable for OOT drivers.

Original-Bug: 42079190
Test: fx test //src/graphics/display
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1159333
Original-Revision: 910a259190fac78212701e5097f6fd2cce8769be
GitOrigin-RevId: d6a983412659ce158ca89ea81871f48e0579b385
Change-Id: Ibc87a00fe8ef6da47ada05329fb11b3e7b5b32bb