[roll] Roll fuchsia [memory] Remove last use of HLCPP in pressure_signaler
This commit remove the last remaining uses of legacy HLCPP bindings in
pressure_signaler that it inherited from memory_monitor; it also
removes pressure_signaler from the visibility allowlist that let it
use HLCPP bindings to begin with. This is a bit of cleanup that should
not impact the behavior of pressure_signaler, and that should make the
code a bit more readable.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1159092
Original-Revision: 1812674d4b20cb94b02f2da55853e8aa3ad11864
GitOrigin-RevId: 4c2da4b25702c0ec3a984936914f600fa098b3ff
Change-Id: Id4cd5a3890d24185ea71813ba8e9fe58e97cf0dc