[roll] Roll fuchsia [kernel] Temporarily remove KASan support on arm
The previous support was found to be making heap allocations in a
pre-heap context*, and was the source of a mystery failure in a
seemingly unrelated and innocuous VM initialization refactor. Given that
and that this support will need to be holistically redesigned to work in
the physboot imminently, the maintainer found it prudent for use to
remove the existing support altogether for now (on the condition of
figuring out and committing to a timeline for the redesign, which is in
the works now).

*Unknown how this was working! The heap allocations were being made
implicitly in arch_asan_early_init() in the usage of CreateSubVmar() and
Map() in asan_map_shadow_for().

Original-Bug: 379891035
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1160133
Original-Revision: 173fcb28b66274c409b297911089daaef64d54ab
GitOrigin-RevId: 326380de9b9f4c2aef848323522a64d78f5693d6
Change-Id: Ibcd251bb47b97f689d4d1b88d2a31af2ad60c938