[roll] Roll fuchsia [owners] Remove chok@google.com from src/tests/end_to_end/lib/perf_utils/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/chok@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1162619
Original-Revision: 020f5f74c892f0850a5163a3f95c8a9ec0a175bc
GitOrigin-RevId: 4d3eed78412315efaa42d7239e709f0ba074ff25
Change-Id: I5e1a9a7e47d2509122f75361aa5d494036184ce8