[roll] Roll fuchsia [magma][platform] Create PlatformDeviceDfv2
This is a DFv2 version of the platform device which makes FIDL
calls instead of banjo.

Use this in the msd-arm-mali driver first. It will also be used
in the msd-vsi-vip once it migrates to DFv2.

This should not have any functional changes to the msd-arm-mali driver.

Original-Bug: b/301005872
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1158115
Original-Revision: 8acacc92bc31ce75aca6663538e969340c736b88
GitOrigin-RevId: a8af6782e9046f2a4f4b0f2a4e7bf04a1e453468
Change-Id: I6aba31a927629b959c860ea91dee9d46a00b826f