[roll] Roll fuchsia [owners] Remove chok@google.com from src/testing/end_to_end/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/chok@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1162773
Original-Revision: be4b35fe984cc4038e5ae7bd845f9627f14b92b9
GitOrigin-RevId: e4c78433422c68dcda5e95acd26d20aa5012d549
Change-Id: Iaa661ae7e32102705eb1d7abe58114e94e2cde81