[roll] Roll fuchsia [memory] Fix logging when not capturing on level changes
When memory_monitor does not capture digests when memory pressure levels
change, we still need to verify that the next capture time is in line
with the capture rate for the level.

This fixes the issue that regular captures are never started when no
capture on level change is configured.

BUG=375462598

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1164632
Original-Revision: ff253663efe8e71f24123a5fb3769b18e01076f7
GitOrigin-RevId: a3eb1d30a15fb8571de3454c870d3f259a752d0a
Change-Id: I7164582d0d21c42380279e5093aa70568e2fbd6f