[roll] Roll fuchsia [paver] Add FindPartitionTableManager
Note: this is a reland of fxrev.dev/1152401, but I've changed the name
of the method to one which (hopefully) makes its purpose and function
more obvious.

Adds fuchsia.paver.Paver/FindPartitionTableManager, which replaces
fuchsia.paver.Paver/UseBlockDevice.

The old API (UseBlockDevice) put the responsibility of finding the
appropriate block device on clients, but this requires clients to have
access to devfs, which is being removed.

The new approach relies on asking the paver service to find the
appropriate block device, and in turn the paver will ask fshost (see
https://fxrev.dev/1151454).

Client will be ported in follow-up CLs.

Original-Bug: 371060853
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1162294
Original-Revision: 0843748936b1161449f3f93fe7137b4d16c6fc89
GitOrigin-RevId: 703ce133a23da49c946c1d86297e0230957e38bb
Change-Id: I18cd1bf2c79d7ea591fb890c4ba965012154ec34