[roll] Roll fuchsia [bazel_sdk] Remove a few un-necessary @fuchsia_sdk load() prefixes.
This is a followup of fxr/1163917 which introduced @rules_fuchsia
as the official name of the repository that contains all SDK rules,
without SDK atom targets in it.

This CL removes a few un-necessary use of @fuchsia_sdk// to load
files that are located in the same rules repository.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1164756
Original-Revision: bddaeae8177c5dcb61be720e7bf8e6e5b2f5980a
GitOrigin-RevId: 4c85edc5f16f98916868d3391b2dc8fb3e7d639a
Change-Id: I72e8117dbaea11592fc8e2dbef412e39a2f64eac