[roll] Roll fuchsia [fidl][build] Add dependency on main library to all backends
Make all FIDL backend targets depend on the main FIDL library target
rather than just the IR target so that additional processing is
guaranteed to be executed. Restrict the visibility of the IR target to
prevent future such issues in the future.

This results in the addition of at least 130 libraries to
`all_fidl_json.txt`.

A number of affected libraries have lint warnings, so linting is
temporarily disabled for them.

Original-Bug: 381123422, 295190178, 381096879
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1163703
Original-Revision: 3066140485b7730db482271e6f7ae41cc5cd6409
GitOrigin-RevId: 6047f3fd970ab62cd2fe4aa53f62c88b8bed5820
Change-Id: I73cd2dae0cc1c2bc035682c520cde09276c422be