[roll] Roll fuchsia [owners] Remove chok@google.com from third_party/pylibs/mypy_extensions/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/chok@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1162616
Original-Revision: 71a7bb043e7d698f642eaeb9d6c3cc681cefed53
GitOrigin-RevId: 0000d2db78b9470f47b2ce76a787330617fce9bc
Change-Id: I38ae39c86d692be7e6c083cfd4bf3d854dd128f8