[roll] Roll fuchsia [owners] Remove chok@google.com from third_party/pylibs/isort/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/chok@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1162779
Original-Revision: befcc0a2aadb225eeb62dea53eb1aff6ba8a52dc
GitOrigin-RevId: 4ff3da4c93f8c81adf4ba3caac1829030b5936ef
Change-Id: I5800419091709ad2b43310f8886d20ca6cc1be50