[roll] Roll fuchsia [build][elfldltl][ld] Elide unnecessary configs
The //build/config/linux:default-pie config was never appropriate
for non-executable targets.  It wasn't much noticed because for a
long time no shared_library() or loadable_module() targets were
ever used in the Linux-targetted toolchains.  It was actively
worked around in a few places.  This corrects that by adding it
to the set_defaults() list only for executable() targets and not
for the other kinds.

For ld-stub, the no_cpp_standard_library and nolibc configs were
superfluous with using -nostdlib, so no need for them.  The
implicit-host-libs config needs to be removed there as well.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1167492
Original-Revision: 493d3530a0cdcb7baad51e824d641e0850728935
GitOrigin-RevId: 3bb0e168fc1e4646bd5282f0c249f05c965b434c
Change-Id: I18a6ac91d11503c405d732f82aebf7f42a8c8a7f