[roll] Roll fuchsia [debugger] Do not assert state when handling exceptions
When handling exceptions, there's really nothing stopping anything else
in the system from coming along and destroying the process or thread
that caused the exception, this is particularly true when debug_agent is
attaching higher in the job hierarchy instead of directly to a process
since debug_agent is not the first thing seeing the exception. Therefore
we should not make assertions on the state of these objects, but
gracefully handle the error cases instead.

This change is not comprehensive for all of DebugAgent, but should
provide a starting off point.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1155933
Original-Revision: a03944c13bdfe46c68cd2b71bb1a97c9ab040b5b
GitOrigin-RevId: f50cdbe7f9c1727b5763a498de07ec4043a78078
Change-Id: Iae391b826c8fe8b3600b71300d561753fc5438aa