[roll] Roll fuchsia [wlan] Ensure PhyManager::create_iface always records an iface creation failure
There was one location where PhyManager::create_iface could fail and not
record the failure as a defect. This change moves the call to
record_defect into create_iface to prevent callers from needing to
remember to call record_defect themselves.

Note that this change will not result in recording a defect in the
previously missed case because PhyManager::add_phy will not insert a phy
and there won't be a place to retain the defect.

Original-Bug: 381908416
Test: Modified existing unit tests to check PhyManager::create_iface
records a defect when a create iface operation fails.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1167173
Original-Revision: 01511b700dbe6f223877098786a70d492b0c228f
GitOrigin-RevId: e1a9aa4ed0fa2f2ae08535b04991eb15667ac836
Change-Id: Iccb5c844402874124bc4506f2bbdfa7a34fe68ee