[roll] Roll fuchsia [starnix] Record inspect data per input file
Currently inspect data is only recorded for an open file and is
discarded once the external program holding a reference to that file
decides to close it. This CL implements behavior that maintains a file's
inspect node in the inspect tree when a file is closed by tying the
lifetime of a input file node to it's parent device. We also introduce a
separate inspect node and inspect data for each separate input file.

Original-Bug: b/365571169
Original-Bug: b/370800942
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1164992
Original-Revision: 2d2660c47ed1d6e1cf5e193c2104e7ebfe36b798
GitOrigin-RevId: f90700961655e663103240eb4c79a4a948afc2cf
Change-Id: Ia970f67065d3b170627b3d382f6217cecda82fe0