[roll] Roll fuchsia [diagnostics] Fix and allow clippy lints.
Addresses lint violations that were allowed to sneak in because of a
build refactor that accidentally allowed all clippy lint categories.

Note that this may somewhat affect benchmark metrics as it refactors
the black_box calls to actually take a value. When black_box is passed
a `()` (as returned by drop()), it's a no-op.

Original-Bug: 381896734
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1167248
Original-Revision: 23e3f9e827d12ddbb113bbfa6968399ce5c649be
GitOrigin-RevId: acc5585abdbc9e8dae32f6a5d38a5a3a5675d429
Change-Id: I8b2624e1867ecb6107ee7310fef0b925d5275bcd