[roll] Roll fuchsia [sestarnix] Write xattr to root nodes if not already labeled
When mounting a filesystem with xattr labelling we expect either
that it is a pre-existing filesystem, in which case the root node
should already have been labeled, or a new filesystem, in which
case the root node needs to have a label written to it.

Since we don't have clean separation of existing vs new for
cases such as "remotefs" mounts on fxfs storage, where the root
will have been created by Fuchsia before handing it off to Starnix,
we approximate the expected behaviour by attempting to write a
label to the root node if none is found when we init_with_dentry().

Original-Bug: 381117111
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1164812
Original-Revision: 817bb17eda010449f6ba72c459a6a9d084c586d7
GitOrigin-RevId: 65bd963f54a413fd1352e018324fd9ff3712cc14
Change-Id: I9e45c5382a30f8088af7fa0754d966af7c77e67b