[roll] Roll fuchsia [netstack3] Introduce SasHandler::get_local_addr_id_for_remote
Add a method to SasHandler to retrieve a strongly-held reference to the
best local address for communicating with a given remote, and implement
the existing get_local_addr_for_remote in terms of it. This will make it
easier for NAT to stash a WeakAddressId, and doesn't affect existing
usages because the AddressId provides easy access to the scalar
IpDeviceAddr.

Original-Bug: 372546692
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1164196
Original-Revision: 70b5aeaf99280ef2ebe54de9e82b2a4adf5fedf5
GitOrigin-RevId: 45280fa81a1a4b8010592414f2533ab677725adb
Change-Id: Ic2e3ab5b75977464ee8ddf13c0ea2fc4547ffd8f