[roll] Roll fuchsia [time] Defines a custom time type for the C++ use of the UTC clock
time_utc is not zx::basic_time<1>.  It was defined like so for
expediency when we split off the boot timeline from the monotonic
timeline.

This change ensures that timekeeper::time_utc is a distinct type, but
still equipped with all the algebraic tooling that is available for the
family of types `zx::basic_time<T>`.

Original-Bug: 376107748
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1167757
Original-Revision: 6c7291885675db660d7e1d9fb7c29bfff2676f59
GitOrigin-RevId: 0997dc64fe4ee96f4b5f041909a19a71eeedc6c9
Change-Id: I62afe6e361988802f9565fc93cc6d1ebdf6e6b8c