[roll] Roll fuchsia [component][test] AddChildRealmFromDecl
Introduces a new API into the component testing FIDL for adding a child
realm that initializes with a compoennt decl. Very similar to the newly
added AddChildRealmFromRelativeUrl but removes the need for the
declaration to be in a file in the pkg, and be dynamically loadable.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1168035
Original-Revision: be0cdae63e8fd77231106e5d872f2f4b6779d666
GitOrigin-RevId: e1d0a7f41567fde1aa8e5f7d05822c006074532d
Change-Id: I5a78886e14525a3386bf3576daddb7100729c0f5