[roll] Roll fuchsia [elf][time] Computes the process' start UTC date-time almost-correctly
The old code interpreted the start time as a point on the boot timeline.
However, this is not strictly correct, as the process start time is
defined as a point on the monotonic timeline.  This was done as a
pragmatic way to expedite migration to boot time clock where appropriate
in Fuchsia, and was temporarily the correct approach, since our
monotonic and boot timelines are identical today.

However, once the two diverge - we will begin to stop monotonic timeline
where appropriate - the approximation will no longer be correct. This is
because the two timelines will drift apart, affecting the accuracy of
that estimation.

The resolution adopted here is to map the process start time through two
clock transformations, one transforming from monotonic to boot, and
another transforming from boot to UTC.

This works acceptably except in one specific case.  Since we do not
prevent the system from going to sleep, we could end up obtaining the
clock transformations before a suspend, but mapping through the
timestamp after the transformation. Since in a suspend the values of the
transformations change, this means we could grow a systematic error
here.

Since the impact of the UTC date-time computation is limited to a very
specific user-observable one-liner in `ffx component show`, and the
fairly small chance that the device sleeps exactly between the two
computations, this should be acceptable in most cases.

Original-Bug: 377525445
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1168240
Original-Revision: f9d55a28fbb09b58811ded11613756e4d8eb9021
GitOrigin-RevId: 34acbaace412e998b14d6e6845defac5b259ef93
Change-Id: Ib894a5fb8a708e9634937a48f4036120f15531ce