[roll] Roll fuchsia [packet-formats] Reinterpret MLDv2/IGMPv3 queries
When operating as an older version listener, if we're capable of
decoding the new version queries we should read them as v1 queries. The
only field that carries a semantic difference there is how to interpret
the response code, which changed in the later versions.

To make the transmutation clearer, introduce methods to reintrepret the
underlying bytes as the older version so netstack3 can operate in strong
older version mode.

Original-Bug: 42071006
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1168175
Original-Revision: d12f524607cef4a2ac58ef282a4bf5ecd7b555ef
GitOrigin-RevId: 489c6eca0aec2685318353bcf526dff957ca076e
Change-Id: I9b59ea60021fb4f83d63b299e3a193812c1ad14d