[roll] Roll fuchsia [sdk][ctf] Use static driver_test_realm shard
.. in hwinfo and diagnostics CTF test components. The previously
configured child components do not resolve because of the missing dep,
which is obscured if the test's root realm is lazily started. When the
root realm is eagerly started, it becomes a hard error that causes the
test to fail.

Further, adding a dep for the previous shard that was included in the
manifests causes other issues because they expect the tests to interact
with the driver_test_realm, which these tests to not.

Because the requirements for these tests are to simply have a statically
configured topography of children, the dependency is changed to the
static driver_test_realm topography offered through the SDK and the
manifest shard is updated accordingly.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1168592
Original-Revision: b557aeb75aafe9f6974b4597c3b55150b6fd53d3
GitOrigin-RevId: 0c4504216038e671e126c97f0250fe64270751c1
Change-Id: Iba3e1c8a771437e2541738a7622e5f0c90c6db1f