[roll] Roll fuchsia [display][engine] Move Engine events to a separate listener protocol.
This change is similar to I01f8a245514897edd94c0e93a813a57fed7154e5
where we moved all events from `Coordinator` to `CoordinatorListener`,

Moving events to a separate protocol gives us more flexibility in
event handling, so that the asynchronous FIDL event handling and
synchronous FIDL Engine methods can coexist in the display coordinator
and engine drivers.

This change adds a `RegisterEngineListener()` method to the `Engine`
protocol for the coordinator to register an event listener. It's
similar to the method of the same name in the banjo protocol, though
the coordinator doesn't need to explicitly deregister the listener
-- the coordinator can close the channel directly.

Test: fx build
Original-Bug: 341156315
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1167932
Original-Revision: 3903ea575bead01a7466048fa6bb0544211f967a
GitOrigin-RevId: e5842cda4515d48fd50925d7d1c477d35c38e01f
Change-Id: I8e4735c9d3e6d417455f7b95ab3c441dc148c1f8